Optimise simplification rules (sort into bins).
